+++ enpt/options/config.py
@@ -118,14 +118,14 @@ config_for_testing_dlr = dict(
                      # Arcachon full tile 3, reprocessed 05/2020
                      # 'ENMAP01-____L1B-DT000400126_20170218T110119Z_003_V000204_20200508T124425Z.zip'
                      )),
-    path_l1b_enmap_image_gapfill=os.path.abspath(
-        os.path.join(path_enptlib, '..', 'tests', 'data', 'EnMAP_Level_1B',
-                     # Alps
-                     # 'ENMAP01-____L1B-DT000000987_20130205T105307Z_001_V000101_20190426T143700Z__rows100-199.zip'
-
-                     # Arcachon
-                     'ENMAP01-____L1B-DT000400126_20170218T110115Z_002_V000204_20200206T182719Z__rows800-899.zip'
-                     )),
+    # path_l1b_enmap_image_gapfill=os.path.abspath(
+    #     os.path.join(path_enptlib, '..', 'tests', 'data', 'EnMAP_Level_1B',
+    #                  # Alps
+    #                  'ENMAP01-____L1B-DT000000987_20130205T105307Z_001_V000101_20190426T143700Z__rows100-199.zip'
+    #
+    #                  # Arcachon
+    #                  # 'ENMAP01-____L1B-DT000400126_20170218T110115Z_002_V000204_20200206T182719Z__rows800-899.zip'
+    #                  )),
     path_dem=os.path.abspath(
         os.path.join(path_enptlib, '..', 'tests', 'data',
                      # Alps
@@ -204,6 +204,10 @@ class EnPTConfig(object):
         :key n_lines_to_append:
             number of lines to be added to the main image [if None, use the whole imgap].
             Requires 'path_l1b_enmap_image_gapfill' to be set.
+
+        :key drop_bad_bands:
+            if set to True (default), the water absorption bands between 1358 and 1453 nm as well
+            as between 1814 and 1961 nm are excluded from processing and will not be contained in the L2A product
 
         :key disable_progress_bars:
             whether to disable all progress bars during processing
@@ -317,6 +321,7 @@ class EnPTConfig(object):
         self.path_l1b_snr_model = self.absPath(gp('path_l1b_snr_model'))
         self.working_dir = self.absPath(gp('working_dir')) or None
         self.n_lines_to_append = gp('n_lines_to_append')
+        self.drop_bad_bands = gp('drop_bad_bands')
         self.disable_progress_bars = gp('disable_progress_bars')
 
         ##################
@@ -378,8 +383,9 @@ class EnPTConfig(object):
         # check if given paths point to existing files
         paths = {k: v for k, v in self.__dict__.items() if k.startswith('path_')}
         for k, fp in paths.items():
-            if fp and not os.path.exists(fp):
-                raise FileNotFoundError("The file path provided at the '%s' parameter does not exist (%s)." % (k, fp))
+            if fp and not os.path.isfile(fp):
+                raise FileNotFoundError("The file path provided at the '%s' parameter does not point to an existing "
+                                        "file (%s)." % (k, fp))
 
         if not self.path_dem:
             warnings.warn('No digital elevation model provided. Note that this may cause uncertainties, e.g., '