+++ HISTORY.rst
@@ -10,6 +10,22 @@ coming soon
   land surfaces later.
 
 
+0.14.1 (2020-09-01)
+-------------------
+
+* Updated installation instructions to include SICOR installation there.
+* Some adaptions to recent changes in py_tools_ds and pyproj.
+
+
+0.14.0 (2020-08-14)
+-------------------
+
+* Implemented new configuration option (parameter 'drop_bad_bands') to leave out certain water absorption bands from
+  processing.
+* Bugfix for allowing directories at filename config parameters.
+* Total runtime is now printed.
+
+
 0.13.2 (2020-08-11)
 -------------------